Create a package for Free and re-organize
